* prelim fix nvarchar(max) on WeightMeasurement

+++ src/DMSpro.OMS.MdmService.Application.Contracts/WeightMeasurements/WeightMeasurementDto.cs
@@ -8,7 +8,7 @@ namespace DMSpro.OMS.MdmService.WeightMeasurements
     {
         public string Code { get; set; }
         public string Name { get; set; }
-        public uint Value { get; set; }
+        public decimal Value { get; set; }
 
         public string ConcurrencyStamp { get; set; }
     }
